Send login methods on user checks and extended pixel lookup.
- Adds a list of UserLogins to ExtendedLookup
- Refactors the ServerUserInfo packet to be a mix of three classes: packets.UserInfo (base), packets.socket.ServerUserInfo (UserInfo with packet type) and packets.http.ExtendedUserInfo (UserInfo with logins, send on user check)
